♻️ Move filters to its own file
