+++ components/Search/index.ts
@@ -1,1 +1,2 @@
 export { SearchResultStats } from "./SearchResultStats"
+export { Filter } from "./Filter"
